check if file exists and is image before deleting; throws exception on failure; using Log

+++ src/main/java/nl/opengeogroep/safetymaps/server/admin/stripes/FotoActionBean.java
@@ -5,10 +5,13 @@
  */
 package nl.opengeogroep.safetymaps.server.admin.stripes;
 
+import java.awt.Image;
 import java.io.BufferedInputStream;
 import java.io.BufferedOutputStream;
 import java.io.File;
 import java.io.FileInputStream;
+import java.io.IOException;
+import static java.lang.Math.log;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
@@ -16,6 +19,7 @@ import java.sql.SQLException;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
+import javax.imageio.ImageIO;
 import javax.naming.NamingException;
 import net.sourceforge.stripes.action.ActionBean;
 import net.sourceforge.stripes.action.ActionBeanContext;
@@ -24,15 +28,20 @@ import net.sourceforge.stripes.action.DontValidate;
 import net.sourceforge.stripes.action.ForwardResolution;
 import net.sourceforge.stripes.action.Resolution;
 import net.sourceforge.stripes.action.SimpleMessage;
+import nl.opengeogroep.safetymaps.server.db.Cfg;
 import nl.opengeogroep.safetymaps.server.db.DB;
 import org.apache.commons.dbutils.handlers.MapListHandler;
+import org.apache.commons.logging.Log;
+import org.apache.commons.logging.LogFactory;
 
 /**
  *
  * @author martijn
  */
 public class FotoActionBean implements ActionBean {
-
+    
+    private static final Log log = LogFactory.getLog(FotoActionBean.class);
+    
     private ActionBeanContext context;
 
     private List<Map<String, Object>> fotos = new ArrayList();
@@ -70,14 +79,13 @@ public class FotoActionBean implements ActionBean {
     @DontValidate
     public Resolution delete() throws Exception {
         String filename = context.getRequest().getParameter("filename");
-        String location = context.getRequest().getParameter("location");
         if (filename != null) {
             try {
-                deleteFromFileSystem(location);
+                deleteFromFileSystem(filename);
                 DB.qr().update("delete from wfs."+TABLE+" where filename ='" + filename + "'");
                 getContext().getMessages().add(new SimpleMessage("Foto verwijderd."));
             } catch (Exception e) {
-                e.printStackTrace();
+                log.error(e);
             }
         } else {
             getContext().getMessages().add(new SimpleMessage("Geen bestandsnaam om te verwijderen!"));
@@ -137,18 +145,24 @@ public class FotoActionBean implements ActionBean {
         return result;
     }
     
-    public void deleteFromFileSystem(String path) {
+    public void deleteFromFileSystem(String filename) throws NamingException, SQLException {
+        String path = Cfg.getSetting("fotofunctie") + filename;
         try {
-
             File file = new File(path);
-
-            if (file.delete()) {
-                System.out.println(file.getName() + " is deleted!");
+            Image image = ImageIO.read(file);
+            if (image != null && file.exists()) {
+                if (file.delete()) {
+                    log.debug(file.getName() + " is deleted!");                    
+                } else {
+                    log.debug("Delete operation failed.");
+                    throw new Error("Delete operation failed");
+                }
             } else {
-                System.out.println("Delete operation failed.");
+                log.debug("The file" + path + "could not be deleted , it is not an image or does nor exist");
+                throw new Error("The file" + path + "could not be deleted , it is not an image or does nor exist");
             }
-        } catch (Exception e) {
-            e.printStackTrace();
+        } catch (IOException e) {
+            log.error(e);
         }
     }
 }
